Move definition of types into DesktopFile
Types will be shared with future DesktopFileWriter class.
