+++ include/linuxdeploy/core/desktopfileentry.h
